moved billingSlider and pricingCard from ui package to billing module
