+++ apps/web/modules/ee/billing/components/pricing-table.tsx
@@ -8,11 +8,11 @@ import { cn } from "@formbricks/lib/cn";
 import { capitalizeFirstLetter } from "@formbricks/lib/utils/strings";
 import { TOrganization, TOrganizationBillingPeriod } from "@formbricks/types/organizations";
 import { Badge } from "@formbricks/ui/components/Badge";
-import { BillingSlider } from "@formbricks/ui/components/BillingSlider";
 import { Button } from "@formbricks/ui/components/Button";
-import { PricingCard } from "@formbricks/ui/components/PricingCard";
 import { isSubscriptionCancelledAction, manageSubscriptionAction, upgradePlanAction } from "../actions";
 import { CLOUD_PRICING_DATA } from "../api/lib/constants";
+import { BillingSlider } from "./billing-slider";
+import { PricingCard } from "./pricing-card";
 
 interface PricingTableProps {
   organization: TOrganization;
